% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_multi.R
\name{separate_multi}
\alias{separate_multi}
\title{separate_multi}
\usage{
separate_multi(data, col, pattern = "[^[:alnum:]]+", into_prefix)
}
\arguments{
\item{data}{data.frame or tibble}

\item{col}{column to be separated}

\item{pattern}{regular expression pattern}

\item{into_prefix}{prefix for new columns}
}
\description{
separate_multi
}
